feat: add Marked instance
